requirements.txt
Django
djangorestframework
git+https://github.com/franzelunico/swapi-python.git

